[fenix] Fix https://github.com/mozilla-mobile/fenix/issues/3122 - Hide the "Bookmark saved" Snackbar when user closes BrowserFragment
Previously the Snackbar was being inflated in the parent of this Fragment so
surviving it being closed.
Tying the Snackbar with the Fragment from which it is shown ensures it will be
effectively hidden whenever the user navigates from the Fragment.

UltraBlame original commit: 425a91db2d7ea5a4cd2f8f38b500210781e1925e