[fenix] For https://github.com/mozilla-mobile/fenix/issues/8803 - review: move forward metric expiration date.
We primarily want to determine if this is a problem area for us to
investigate rather than a long term measurement to keep so we should set
the expiration date accordingly. Furthermore, this code executes before
crash reporting is init so it's ideal to remove it sooner rather than
later.

UltraBlame original commit: 7c3e44924ab1536f5849706355f4459fa63aa8b7