[fenix] For https://github.com/mozilla-mobile/fenix/issues/8399: Hide keyboard to prevent content resizes
This hides the keyboard after committing a URL in the
Toolbar right before we navigate from the SearchFragment
to the BrowserFragment. If the BrowserFragment is being
displayed before the keyboard is gone an expensive
resize of the engine view (content) is triggered when the
keyboard finally goes away. This is to prevent that.

UltraBlame original commit: 52973facd6291e42ae472aa6261115a7e343b3a7