[fenix] For issue https://github.com/mozilla-mobile/fenix/issues/10727 - Logins auth redirect
Added an extension function that pops the backstack of the fragment so the user is redirected to the Logins and passwords screen.
This is done to force the user to re-authenticate if he wants to re-enter the saved logins flow.

Additionally, some UI elements are being hidden on each Fragment since they were remained visible on the Logins and passwords screen's UI.

UltraBlame original commit: 265ae87ddaca4cff17e59fdf9efeb1a5a90ddc0e