[fenix] Issue https://github.com/mozilla-mobile/fenix/issues/11333: Set ContentBlocking settings directly on GeckoRuntime
We set the ContentBlockingSettings directly on the GeckoRuntime now to
improve the startup of the engine.

This change has requirements from Android Components and GeckoView, so
we would only see the full perf benefits in Nightly as the changes ride
the train, although we might start to see some of them as we're updating
the GeckoProvider for the `geckoBeta` variant as well.

Co-authored-by: Arturo Mejia <arturomejiamarmolgmail.com>

UltraBlame original commit: 79632b91bb40d05534fa70dd6024c68dd1280f59