[fenix] For https://github.com/mozilla-mobile/fenix/issues/13156, https://github.com/mozilla-mobile/fenix/issues/13280 - Use payloads to bind bookmark viewholders.
Change the DiffUtil callback for bookmarks to use the generated equals()
method instead of only checking the title and url fields. This prevents
the BookmarkNode in our state from getting out of sync with the
BookmarkNode the viewholder is bound to.

UltraBlame original commit: ce1b1eceef42921886b8611e524a6dfd0dd01978