[fenix] No issue: remove inline from StrictModeManager.resetAfter.
I believe this was a premature optimization: it's unclear that it would
actually improve performance and yet  we had to add several weird
workarounds to make it work that broke encapsulation. It's also
possible it would be worse for the APK size because of excessive
inlining.

UltraBlame original commit: a4a716545061ce169fb7aac1cd7b8b15da3a3996