[fenix] Close https://github.com/mozilla-mobile/fenix/issues/18862: Add multi-select banner to tabs tray  (https://github.com/mozilla-mobile/fenix/pull/18932)
* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Add new addBookmark BookmarksUseCase

* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Add class for state binding features

* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Add delete multiple tabs to tray interactor

* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Add new actions to navigation interactor

* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Enable select mode from main tray menu

* Issue https://github.com/mozilla-mobile/fenix/issues/18862: Add menu when in select mode

* Close https://github.com/mozilla-mobile/fenix/issues/18862: Add multi-select banner to tabs tray

* Close https://github.com/mozilla-mobile/fenix/issues/18862: Add select support for handle UI

We apply various layout changes to the "handle" UI in the tabs tray when
switching modes. It isn't quite clear to my, why we do this, if it's
really needed to meet the end result, and if there is a better way.

For now, we're simplying moving over that logic that we can re-evaluate
at a later time.

UltraBlame original commit: faf71ad02b0c1c0faa94fe01045b6be00b4c825a