[fenix] Issue https://github.com/mozilla-mobile/fenix/issues/19175: Fix SyncTabs list not updating on changes
The main cause for this is that the `LifecycleProvider` needs to be set
to `State.RESUMED` to avoid the account manager's internal
`ObserverRegistry` from putting the UI observers into the paused state.

The rest of the changes is to rely the internal (safe) logic to
correctly sync and then update the tabs list.

UltraBlame original commit: 10017565d15ae0c2dead548375e022b12324dee9