[fenix] Issue https://github.com/mozilla-mobile/fenix/issues/21236: Fixes empty tray visibility logic
This is a bug we noticed after landing search term grouping.

An adapter can submit an empty list of items to the `ConcatAdapter`
early. This has the side-effect of triggering our `observeFirstInsert`
too soon and therefore updating the visibility to show the empty tray
placeholder and never switches back.

Our solution is to keep a constant observer on the adapter so we can
perform the visibility check on every insert/remove.

Co-authored-by: Roger Yang <royangmozilla.com>

UltraBlame original commit: 6d2a582ce9f0c9fa77f89e6405595f4b51827d7f